Remove vim-fugitive submodule
I wasn't really using it, and it behaves as though
GIT_DISCOVERY_ACROSS_FILESYSTEM is set, which means it tries to stat
some directories that take a long time to browse, such as /home. That
means if I run vim or view in a non-Git folder, the whole process hangs
indefinitely.