Reverse order of building PATH
By building the path through appending instead of prepending, any items
that weren't enumerated in .PATHrc will bubble up to the front of the
path instead of ending up at the end. That fixes a problem with RVM,
which complains if the RVM folders aren't at the start of the path.
Currently, we don't add those folders in .PATHrc because they aren't
exactly known.

Since .PATHrc is now sourced instead of just read, we could put some
code in there to discover and echo the RVM folders, but for now, it's
enough to just change the order we use to build the path.