Set tmux default-terminal globally
Certain tmux versions are sensitive to whether this is set globally or
per-session. It's a session variable, and maybe it really should be set
as a session variable in case I connect to the same tmux server from
differently capable terminals at the same time, but it's probably fine
to hard-code the terminal type to be 256-color-capable.