Slightly update vim-signature and vim-signify
Vim-signature isn't technically compatible with vim-signify; it prefers
vim-gitgutter instead. A January 2017 update to vim-signify breaks
whatever compatibility there was, so the update to the versions in this
repo avoid that commit for now.

Since we don't use CVS at work anymore, the Git-only limitation in
vim-gitgutter isn't necessarily the dealbreaker it used to be. There's
no time right now to switch plugins, though.