Consume changes from local gitconfig
Somehow, the .gitconfig on my main system got "disconnected" from the
dotfiles copy and was a standalone file. Some changes had been made to
it outside the dotfiles repo.

Changes related to git-diff output highlighting, and the removal of the
git-ls-files alias in favor of the standalone command in dotfiles/bin.